requirements.txt
asgiref==3.3.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.5
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2020.5
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
uuid==1.30
whitenoise==5.2.0
